+++ esdm/esdm_config.c
@@ -36,6 +36,7 @@ struct esdm_config {
 	uint32_t esdm_es_irq_entropy_rate_bits;
 	uint32_t esdm_es_krng_entropy_rate_bits;
 	uint32_t esdm_es_sched_entropy_rate_bits;
+	uint32_t esdm_es_hwrand_entropy_rate_bits;
 	uint32_t esdm_drng_max_wo_reseed;
 	uint32_t esdm_max_nodes;
 	enum esdm_config_force_fips force_fips;
@@ -74,6 +75,11 @@ static struct esdm_config esdm_config = {
 	 * See documentation of ESDM_SCHED_ENTROPY_RATE
 	 */
 	.esdm_es_sched_entropy_rate_bits = ESDM_SCHED_ENTROPY_RATE,
+
+	/*
+	 * See documentation of ESDM_HWRAND_ENTROPY_RATE
+	 */
+	.esdm_es_hwrand_entropy_rate_bits = ESDM_HWRAND_ENTROPY_RATE,
 
 	/*
 	 * See documentation of ESDM_DRNG_MAX_WITHOUT_RESEED.
@@ -184,6 +190,21 @@ void esdm_config_es_sched_entropy_rate_set(uint32_t ent)
 }
 
 DSO_PUBLIC
+uint32_t esdm_config_es_hwrand_entropy_rate(void)
+{
+	return esdm_config.esdm_es_hwrand_entropy_rate_bits;
+}
+
+DSO_PUBLIC
+void esdm_config_es_hwrand_entropy_rate_set(uint32_t ent)
+{
+	uint32_t val = esdm_config_entropy_rate_max(ent);
+
+	esdm_config.esdm_es_hwrand_entropy_rate_bits = val;
+	esdm_es_add_entropy();
+}
+
+DSO_PUBLIC
 uint32_t esdm_config_drng_max_wo_reseed(void)
 {
 	/* If DRNG operated without proper reseed for too long, block ESDM */
@@ -265,6 +286,10 @@ int esdm_config_init(void)
 		esdm_config_entropy_rate_max(
 			esdm_config.esdm_es_irq_entropy_rate_bits);
 	complete_entropy_rate += esdm_config.esdm_es_irq_entropy_rate_bits;
+	esdm_config.esdm_es_hwrand_entropy_rate_bits =
+		esdm_config_entropy_rate_max(
+			esdm_config.esdm_es_hwrand_entropy_rate_bits);
+	complete_entropy_rate += esdm_config.esdm_es_hwrand_entropy_rate_bits;
 
 	if (complete_entropy_rate < esdm_get_seed_entropy_osr(true)) {
 		logger(LOGGER_WARN, LOGGER_C_ES,