+++ esdm/esdm_config.h
@@ -119,6 +119,23 @@ void esdm_config_es_sched_entropy_rate_set(uint32_t ent);
 uint32_t esdm_config_es_sched_entropy_rate(void);
 
 /**
+ * @brief /dev/hwrng ES configuration: set the entropy rate
+ *
+ * NOTE: The ESDM ensures that the entropy rate cannot be set to a value larger
+ *	 than the security strength of the the applied DRNG.
+ *
+ * @param ent [in] Entropy rate in bits.
+ */
+void esdm_config_es_hwrand_entropy_rate_set(uint32_t ent);
+
+/**
+ * @brief /dev/hwrng ES configuration: get the entropy rate
+ *
+ * @return Entropy rate in bits
+ */
+uint32_t esdm_config_es_hwrand_entropy_rate(void);
+
+/**
  * @brief DRNG Manager configuration: get maximum value without successful
  *	  reseed
  *