+++ CHANGES.md
@@ -1,6 +1,8 @@
 Changes 0.5.0-prerelease:
 * Linux kernel entropy feeder is now always enabled
 
+* Add Linux /dev/hwrng entropy source
+
 Changes 0.4.0:
 * Start CUSE daemons independently from ESDM server